Now Riding the Tank is more "Normal"
Riding the Tank more Normal - you're set to third person view if not
already in it. You can switch to between the two third person views.
Your player model doesn't render so you simply interact with the APC
without you getting in the way. Users assume the player is safe in the
tank operating it.